Dual linear + polynomial timer for #604 (#726)
* [FIX]: fix request error bug limiting requests

* [FEAT]: add dual poly. backoff + linear timer if request limit hit

Add a dual polynomial + linear timer to allow for optimal fine grain
searching of twitter timeout value. A new arg `min_wait_time` allows
users to to do one of three things; Firstly, if they know a value that
is optimal already then can input it directly allowing the timer to
stricty be a linear timer. Secondly, if Twitter suddenly changes their
timeout limit and the value no longer satifies the required amount,
the timer will decide the higher valued option between the polynomial
timer's and linear timer's next value so the application still
functions. Third, as the linear timers min_wait_time arg goes to zero
the timer becomes strictly a polynomial backoff timer that allows the
user versatility and less likely to explode in the case of a change in
Twitter request limit timeout.

* [FIX]: add args to config.py

* [REFACT]: minor changes to run.py

* [REFACT]: small change to arg case